feat: refactor user information display into separate UserInfo component
